requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-text-splitters
langchain-google-genai
faiss-cpu
PyPDF2
sentence-transformers